use export_test pattern
